Analyze current offering on the Chrome Web Store
Analyze current offering on the Chrome Web Store.
Works with all apps, extensions, themes, and games on the Chrome Web Store. Must be currently offered in the store.
Gives a rating and can provide a greater insight into what data is requested as well as permissions, Content Security Policy, and more.